Rename the Director KickScript skill to ScriptKick (#1022)
