+++ module/skill/ScriptKick/CMakeLists.txt
