+++ module/skill/ScriptKick/data/config/ScriptKick.yaml
