+++ roles/test/directormotion.role
@@ -9,7 +9,7 @@ nuclear_role(
   actuation::Kinematics
   actuation::Servos
   # Skills
-  skill::KickScript
+  skill::ScriptKick
   skill::GetUp
   # Planners
   planning::GetUpPlanner
